LayoutTests/inspector/css/getAllStyleSheets.html

<!DOCTYPE html>
<html>
<head>
<script src="../../http/tests/inspector/resources/inspector-test.js"></script>
<link rel="stylesheet" href="resources/external.css">
<style>body { color: blue; }</style>
<script>
var style = document.createElement("style");
style.textContent = "body { color: gray; }";
document.head.appendChild(style);
</script>
</script>
<script>
function test()
{
    function sanitizeURL(url) {
        return url.replace(/^.*?LayoutTests\//, "");
    }

    WebInspector.cssStyleManager.fetchStyleSheetsIfNeeded();
    InspectorBackend.runAfterPendingDispatches(function() {
        let styleSheets = WebInspector.cssStyleManager.styleSheets;
        InspectorTest.expectThat(styleSheets.length === 3, "There should be three stylesheets.");

        for (let styleSheet of styleSheets) {
            InspectorTest.log("");
            InspectorTest.log(`URL: ${styleSheet.isInlineStyleTag() ? "<style> on " : ""}${sanitizeURL(styleSheet.url)}`);
            InspectorTest.log(`OFFSET: (${styleSheet.startLineNumber}, ${styleSheet.startColumnNumber})`);
            InspectorTest.expectThat(styleSheet.parentFrame, "Stylesheet should have a frame.");
        }

        InspectorTest.completeTest();
    });
}
</script>
</head>
<body onload="runTest()">
<p>Test CSS.getAllStyleSheets returns expected information for different style sheet sources.</p>
</body>
</html>
